extract common class to base module
