+++ inlong-sort/sort-connectors/postgres-cdc/src/main/java/org/apache/inlong/sort/cdc/postgres/table/PostgreSQLDeserializationConverterFactory.java
@@ -17,14 +17,14 @@
 
 package org.apache.inlong.sort.cdc.postgres.table;
 
-import com.ververica.cdc.debezium.table.DeserializationRuntimeConverter;
-import com.ververica.cdc.debezium.table.DeserializationRuntimeConverterFactory;
 import io.debezium.data.SpecialValueDecimal;
 import io.debezium.data.VariableScaleDecimal;
 import java.math.BigDecimal;
 import java.time.ZoneId;
 import java.util.Optional;
 import org.apache.flink.table.types.logical.LogicalType;
+import org.apache.inlong.sort.base.debezium.table.DeserializationRuntimeConverter;
+import org.apache.inlong.sort.base.debezium.table.DeserializationRuntimeConverterFactory;
 import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.Struct;
